add asbites+astext, remove asdecodedhex, fix crash
asbites(num_bits) returns a bytearray. num_bits must be a multiple of 8.

astext() is a drop-in replacement for asdecodedhex().
By default text output stops after the first null byte.
Default encoding is 'latin-1' for robustness, as it is more permissive and clearer than other encodings in case of unexpected input.

This also fixes a potential crash in fromhex() that would happen if hex() returned a odd number of nibbles depending on the content.